Modifying pr and release workflows
Decoupling both pr and release workflows as independent and leaving pr only for pull request on master.
The content of pr has been added to release in case any user push something onto master without a previous pull request.
The reason of this duplication of code is because a workflow can't be referenced from another workflow.

Another modification on the release workflow is the workflow itself. Now every job depends on the previous one to be
executed, creating a chained shape. Release will be only applicable when changes are made under charts or templates directories.